Always return a promise
To stop chained .then / .catch in the mocha runner from throwing.

The return of rejectFn gets returned at runnable.js:127, but was
previously always undefined.

mocha.js:97 assumes that it can chain off this value and so any test
that throws is handled, but then causes:
    TypeError: Cannot read property 'then' of undefined
which crashes the whole test process.

Returning the rejected promise avoids this problem by allowing the
.catch in mocha.js to be chained and handle the error.